% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infMaxwell.R
\name{fac.os.TLs}
\alias{fac.os.TLs}
\title{One-sided tolerance interval of two-parameter Maxwell distributions}
\usage{
fac.os.TLs(nr = 10^5, n, p, cl)
}
\arguments{
\item{nr}{numeric vector length one, number of simulation runs, defaults to 100,000.}

\item{n}{a numeric vector, the number of observation in each sample}

\item{p}{a numeric vector of probabilities}

\item{cl}{confidence level as in decimal, numeric value between 0 and 1.}
}
\description{
This function computes (p, cl) one-sided tolerance factors in order to calculate one sided tolerance intervals.
}
\examples{
# fac.os.TLs(10^5, n = 10, p = .9, cl = .95)

}
\keyword{fac.os.TLs}
