% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infMaxwell.R
\name{dMaxwell}
\alias{dMaxwell}
\alias{pMaxwell}
\alias{qMaxwell}
\alias{rMaxwell}
\title{Two-parameter Maxwell-Boltzmann Distribution}
\usage{
dMaxwell(x, mu = 0, sig = 1)

pMaxwell(q, mu = 0, sig = 1, lower.tail = TRUE)

qMaxwell(p, mu = 0, sig = 1, lower.tail = TRUE)

rMaxwell(n, mu = 0, sig = 1)
}
\arguments{
\item{x, q}{numeric vector of qunatiles.}

\item{mu}{vector of location parameter. Default is 0.}

\item{sig}{vector of scale parameter. Default is 1.}

\item{lower.tail}{logical; If TRUE(default), probabilities are \eqn{P[ X \leq x]}, otherwise, \eqn{P[X \gt x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations; If \code{length(n)>1}, the length is taken to be the number required.}
}
\value{
\code{dMaxwell} gives the density, \code{pMaxwell} gives the cumulative distribution function, \code{qMaxwell} gives quantile functions, and \code{rMaxwell} generates random numbers from the two-parameter Maxwell distribution.

The length of the result is determined by \code{n} for \code{rMaxwell},  and is the maximum of the lengths of the numerical arguments for the other functions.

The numerical arguments other than \code{n} are recycles to the length of the result. Only the first element of the logical vector is being used.

In case of \code{pMaxwell} and and \code{qMaxwell}, if there is no specification about the logical argument \code{lower.tail}, they assume it's \code{True} by default, which means probabilities are \eqn{P[X \leq x]}.
}
\description{
Density, distribution function, quantile function, and random generation for the two-parameter Maxwell-Boltzmann distribution. This probability distribution has a location parameter \eqn{\mu}, and a scale parameter \eqn{\sigma}.
}
\details{
If location and scale parameter are not specified, above functions will assume default value as 0 and 1, respectively.

The probability density function (pdf) of two parameter Maxwell-Boltzmann distribution (popularly known as Maxwell distribution) is given by
\deqn{f(x|\mu,\sigma) = \frac{4}{\sigma\Gamma(1/2)}\left(\frac{x-\mu}{\sigma}\right)^2\exp\left\{-\left(\frac{x-\mu}{\sigma}\right)^2\right\}, \quad x>\mu, \ \sigma >0}
Where \eqn{\mu} and \eqn{\sigma} are the location and scale parameter respectively.

See \href{ https://doi.org/10.1007/s42519-022-00270-y}{Statistical Intervals for Maxwell Distributions} for more details.
}
\examples{
#For calculating densities
dMaxwell(2,1,2)
dMaxwell(5, 0, 5)

# In order to plot the pdf with a fixed location parameter (say 0) with varying scale parameter
x<-seq(0,5,length=100)
sigmas<-c(.5,1,2)
plot(x,dMaxwell(x,0,.5), type="l",col="blue",xlab="x",ylab="Probability density f(x)")
lines(x,dMaxwell(x,0,1),col="red")
lines(x,dMaxwell(x,0,2),col="green")
legend("topright", legend = paste("\u03C3 =", sigmas), col = c("blue", "red", "green"), lty = 1, lwd = 2)

# For calculating cumulative probabilities
pMaxwell(2,1,2)
pMaxwell(2,0,2,lower.tail=F)

# For calculating quantiles
qMaxwell(.5,1,2)
qMaxwell(.5,0,2,lower.tail=F)

# For generating random numbers
rMaxwell(5,1,2)
rMaxwell(25,0,2)

}
\references{
Chowdhury, F. and Krishnamoorthy, K. (2022). Statistical intervals for Maxwell distributions. Journal of Statistical Theory and Practice. 16, 45.
}
\keyword{dMaxwell}
\keyword{pMaxwell}
\keyword{qMaxwell}
\keyword{rMaxwell}
