% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infMaxwell.R
\name{MEs}
\alias{MEs}
\title{Moment Estimates (MEs) calculating function of two-parameter Maxwell distributions}
\usage{
MEs(x)
}
\arguments{
\item{x}{a numeric vector of data input.}
}
\description{
This function allows you to calculate the MEs of location & scale parameter respectively
}
\examples{
 # MEs(x)
}
\keyword{MEs}
