% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infMaxwell.R
\name{MLEs}
\alias{MLEs}
\title{Maximum likelihood estimates (MLEs) calculating function of two-parameter Maxwell distributions}
\usage{
MLEs(x)
}
\arguments{
\item{x}{a numeric vector of data input.}
}
\description{
This function allows you to calculate the MLEs of location & scale parameter respectively
}
\examples{
 # MLEs(x)
}
\keyword{MLEs}
