% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infMaxwell.R
\name{one.sided.MLE.TL}
\alias{one.sided.MLE.TL}
\title{One-sided tolerance interval (lower and upper) of two-parameter Maxwell distributions}
\usage{
one.sided.MLE.TL(x, nr, p, cl, tails = T)
}
\arguments{
\item{x}{numeric vector of data input}

\item{nr}{numeric vector length one, number of simulation runs, defaults to 100,000.}

\item{p}{a numeric vector of probabilities}

\item{cl}{confidence level as in decimal, numeric value between 0 and 1.}

\item{tails}{logical; If TRUE (default), function calculate lower tolerance limit. Otherwise, function calculate upper tolerance limit.}
}
\description{
This function computes (p, cl) one-sided lower tolerance limit (LTL) and one sided upper tolerance limit (UTL).
}
\examples{
# x<-c(135, 98, 114, 137, 138, 144, 99, 93, 115, 106, 132, 122, 94, 98, 127,
# 122, 102, 133, 114, 120, 93, 126, 119, 104, 119, 114, 125, 107, 98, 117, 111,
# 106, 108, 127, 126, 135, 112, 94, 127, 99, 120, 120, 121, 122, 96, 109, 123, 105)
# lower tolerance limit
# one.sided.MLE.TL(x,10^5,.9,.95,T)
# upper tolerance limit
# one.sided.MLE.TL(x,10^5,.9,.95,F)

}
\keyword{one.sided.MLE.TL}
