% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infMaxwell.R
\name{perc.ci.mean.MEs}
\alias{perc.ci.mean.MEs}
\title{Confidence intervals (MEs based) for two-parameter Maxwell distributions}
\usage{
perc.ci.mean.MEs(cl, n, nr = 10^5)
}
\arguments{
\item{cl}{confidence level as in decimal, numeric value between 0 and 1.}

\item{n}{a numeric vector, the number of observation in each sample}

\item{nr}{numeric vector length one, number of simulation runs, defaults to 100,000.}
}
\description{
This function computes the required percentiles to compute confidence interval for the mean (moment estimator based) of Maxwell distributions.
}
\examples{
 # perc.ci.mean.MEs(.95,5,10^4)
}
\keyword{perc.ci.mean}
