% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infMaxwell.R
\name{perc.ci.mean.MLEs}
\alias{perc.ci.mean.MLEs}
\title{Confidence intervals (MLEs based) for two-parameter Maxwell distributions}
\usage{
perc.ci.mean.MLEs(cl, n, nr = 10^5)
}
\arguments{
\item{cl}{confidence level as in decimal}

\item{n}{a numeric vector, the number of observation in each sample}

\item{nr}{numeric vector length one, number of simulation runs, defaults to 100,000.}
}
\description{
This function computes the required percentiles to compute confidence interval for the mean (MLEs based) of Maxwell distributions.
}
\examples{
 # perc.ci.mean.MLEs(.95, 10,10^5)
}
\keyword{perc.ci.mean.MLEs}
