% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infMaxwell.R
\name{PI.fac}
\alias{PI.fac}
\title{Prediction interval (PIs) of two-parameter Maxwell distributions}
\usage{
PI.fac(nr, n, m, cl)
}
\arguments{
\item{nr}{numeric vector length one, number of simulation runs, defaults to 100,000.}

\item{n}{a numeric vector, the number of observation in each sample}

\item{m}{a numeric vector, the mean of future sample size}

\item{cl}{confidence level as in decimal, numeric value between 0 and 1.}
}
\description{
This function computes the required percentiles to compute prediction intervals of Maxwell distributions.
}
\examples{
 # PI.fac(10^5, n = 45, m = 5, cl = .95)
}
\keyword{PI.fac}
