% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infMaxwell.R
\name{Q_Q_plot}
\alias{Q_Q_plot}
\title{Q-Q plot function for two-parameter Maxwell distributions}
\usage{
Q_Q_plot(x)
}
\arguments{
\item{x}{a numeric vector of data input, should be with length more than one.}
}
\description{
This Q-Q plot function will calculate theoretical Maxwell quantiles based on moment estimates and plot it against sample quantiles in order check goodness of fit.
}
\examples{
 # Q_Q_plot(x)
}
\keyword{Q_Q_plot}
