% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infMaxwell.R
\name{TFs.Maxwell}
\alias{TFs.Maxwell}
\title{Two-sided or equal-tailed tolerance factor}
\usage{
TFs.Maxwell(nr = 10^5, n, p, cl, tails)
}
\arguments{
\item{nr}{numeric vector length one, number of simulation runs, defaults to 100,000.}

\item{n}{numeric vector, number of observation in each sample}

\item{p}{a numeric vector of probabilities}

\item{cl}{confidence level as in decimal, numeric value between 0 and 1.}

\item{tails}{logical; If TRUE, function calculate Equal-tailed tolerance factors. Otherwise, function calculate two-sided tolerance factor.}
}
\description{
This function computes (p, cl) two-sided or equal-tailed tolerance factor
}
\examples{
For two-sided tolerence interval
# TFs.Maxwell(10^5, n = 20, p = .9, .95, tails = F)
For equal-tailed tolerance interval
# TFs.Maxwell(10^5, n = 20, p = .9, .95, tails = T)

}
\keyword{TFs.Maxwell}
